+++ src/pages/login/types.ts
@@ -1,0 +1,3 @@
+export interface ILoginFormProps {
+  setRegisterState?: Function;
+}
